Switch to using greenthreads
At the moment, we're running event listeners as well as workers
processing ports in native threads, while RPC callbacks are to be
executed in greenthreads (yet the threading module is not monkey
patched).

This change attempts to simplify this concurrency model, switching
to using greenthreads. Blocking WMI operations are now transparently
handled by PyMI using eventlet.tpool, starting with PyMI 1.0.4.

Change-Id: Ic34b7f8d53bc51324098e5e6364d303d1eab130d